Remove closure way in Forecast windguru service requests.
