+++ Forescoop/Classes/ForecastWindguruService+Rest.swift
@@ -1,42 +1,49 @@
 //
-//  ForecastWindguruService+Rest.swift
+//  ForecastWindguruService+AsyncAwait.swift
 //  Forescoop
 //
-//  Created by javierfuchs on 7/12/17.
-//
+//  Created by fox on 01/06/2023.
 //
 
-/*
 import Foundation
 
 public extension ForecastWindguruService {
+
+    // user = api(query: routine.user,
+    //        errorCode: err.user.rawValue,
+    //       parameters: [parameter.username,
+    //                   parameter.password]) // not user/pass (anonymous)
     
+    func login(withUsername username: String?,
+               password: String?) async throws -> User? {
+        try await request(tokens: [Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .user)
+    }
+    
+    // searchSpots = api(query: routine.search_spots,
+    //               errorCode: err.search_spots.rawValue,
+    //              parameters: [parameter.search,
+    //                           parameter.opt])
+    func searchSpots(byLocation location: String) async throws -> SpotResult? {
+        try await request(tokens: [Definition.service.api.parameter.search:
+                                    location.addingPercentEncoding(withAllowedCharacters: .urlHostAllowed)],
+                          api: .searchSpots)
+    }
+
     // forecast = api(query: routine.forecast,
     //            errorCode: err.forecast.rawValue,
     //           parameters: [parameter.id_spot,
     //                        parameter.id_model,
     //                        parameter.no_wave])
-    /
-    func forecast(bySpotId spotId: String,
-                  model modelId: String? = nil,
-                  failure: @escaping FailureType,
-                  success: @escaping (_ spotForecast: SpotForecast?) -> Void)
-    {
-        let tokens = [Definition.service.api.parameter.id_model : modelId ?? Definition.defaultModel,
-                      Definition.service.api.parameter.id_spot : spotId]
-        let api = Definition.service.api.forecast
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
-    }
     
+    func forecast(bySpotId spotId: String,
+                  model modelId: String? = nil) async throws -> SpotForecast? {
+        try await request(tokens: [Definition.service.api.parameter.id_model: modelId ?? Definition.defaultModel,
+                                   Definition.service.api.parameter.id_spot: spotId],
+                          api: .forecast)
+    }
+
     //  wforecast = api(query: routine.wforecast,
     //              errorCode: err.wforecast.rawValue,
     //             parameters: [parameter.username,  // pro users
@@ -48,49 +55,21 @@ public extension ForecastWindguruService {
     func wforecast(bySpotId spotId: String,
                    model modelId: String? = nil,
                    username: String? = nil,
-                   password: String? = nil,
-                   failure: @escaping FailureType,
-                   success: @escaping (_ spotForecast: WSpotForecast?) -> Void)
-    {
-        var tokens = [Definition.service.api.parameter.id_model : modelId ?? Definition.defaultModel,
-                      Definition.service.api.parameter.id_spot : spotId]
-        if let username = username, username.count > 0,
-           let password = password, password.count > 0 {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password]
-        }
-        let api = Definition.service.api.wforecast
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                   password: String? = nil) async throws -> WSpotForecast? {
+
+        return try await request(tokens: [Definition.service.api.parameter.id_model: modelId ?? Definition.defaultModel,
+                                          Definition.service.api.parameter.id_spot: spotId,
+                                          Definition.service.api.parameter.username: username,
+                                          Definition.service.api.parameter.password: password],
+                                 api: .wforecast)
     }
     
     // spotInfo = api(query: routine.spot,
     //            errorCode: err.spot.rawValue,
     //           parameters: [parameter.id_spot])
-    func spotInfo(bySpotId spotId: String,
-                  failure: @escaping FailureType,
-                  success: @escaping (_ spotInfo: SpotInfo?) -> Void)
-    {
-        let tokens = [Definition.service.api.parameter.id_spot : spotId]
-        let api = Definition.service.api.spotInfo
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+    func spotInfo(bySpotId spotId: String) async throws -> SpotInfo? {
+        try await request(tokens: [Definition.service.api.parameter.id_spot : spotId],
+                          api: .spotInfo)
     }
     
     // customSpots = api(query: routine.c_spots,
@@ -99,26 +78,10 @@ public extension ForecastWindguruService {
     //                           parameter.password,
     //                           parameter.opt])  // opt=simple (optional)
     func customSpots(withUsername username: String?,
-                     password: String?,
-                     failure: @escaping FailureType,
-                     success: @escaping (_ spots: SpotResult?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let username = username, username.count > 0,
-           let password = password, password.count > 0 {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password]
-        }
-        let api = Definition.service.api.customSpots
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                     password: String?) async throws -> SpotResult? {
+        try await request(tokens: [Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .customSpots)
     }
     
     
@@ -129,26 +92,10 @@ public extension ForecastWindguruService {
     //                             parameter.password,
     //                             parameter.opt]) // opt=simple (optional)
     func favoriteSpots(withUsername username: String?,
-                       password: String?,
-                       failure: @escaping FailureType,
-                       success: @escaping (_ spots: SpotResult?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let username = username,
-           let password = password {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password]
-        }
-        let api = Definition.service.api.favoriteSpots
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                       password: String?) async throws -> SpotResult? {
+        try await request(tokens: [Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .favoriteSpots)
     }
     
     // setSpots = api(query: routine.sets,
@@ -157,26 +104,10 @@ public extension ForecastWindguruService {
     //                        parameter.password])
     //
     func setSpots(withUsername username: String?,
-                  password: String?,
-                  failure: @escaping FailureType,
-                  success: @escaping (_ sets: SetResult?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let username = username,
-           let password = password {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password]
-        }
-        let api = Definition.service.api.setSpots
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                  password: String?) async throws -> SetResult? {
+        try await request(tokens: [Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .setSpots)
     }
     
     // addSetSpots = api(query: routine.set_spots,
@@ -187,32 +118,13 @@ public extension ForecastWindguruService {
     //                           parameter.opt])
     func addSetSpots(withSetId setId: String?,
                      username: String?,
-                     password: String?,
-                     failure: @escaping FailureType,
-                     success: @escaping (_ spots: SpotResult?) -> Void)
-    {
-        var tokens : [String: String?] = [:]
-        if let username = username,
-           let password = password,
-           let id = setId {
-            tokens = [Definition.service.api.parameter.id_set : id,
-                      Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password]
-        }
-        let api = Definition.service.api.addSetSpots
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                     password: String?) async throws -> SpotResult? {
+        try await request(tokens: [Definition.service.api.parameter.id_set: setId,
+                                   Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .addSetSpots)
     }
-    
-    
-    
+
     // addFavoriteSpot = api(query: routine.add_f_spot,
     //                   errorCode: err.add_f_spot.rawValue,
     //                  parameters: [parameter.username,
@@ -220,28 +132,11 @@ public extension ForecastWindguruService {
     //                               parameter.id_spot])
     func addFavoriteSpot(withSpotId spotId: String?,
                          username: String?,
-                         password: String?,
-                         failure: @escaping FailureType,
-                         success: @escaping (_ response: WGSuccess?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let username = username,
-           let password = password,
-           let spotId = spotId {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password,
-                      Definition.service.api.parameter.id_spot : spotId]
-        }
-        let api = Definition.service.api.addFavoriteSpot
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                         password: String?) async throws -> WGSuccess? {
+        try await request(tokens: [Definition.service.api.parameter.id_spot: spotId,
+                                   Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .addFavoriteSpot)
     }
     
     // remove_f_spot = api(query: routine.remove_f_spot,
@@ -251,28 +146,11 @@ public extension ForecastWindguruService {
     //                             parameter.id_spot])
     func removeFavoriteSpot(withSpotId spotId: String?,
                             username: String?,
-                            password: String?,
-                            failure: @escaping FailureType,
-                            success: @escaping (_ response: WGSuccess?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let username = username,
-           let password = password,
-           let spotId = spotId {
-            tokens = [Definition.service.api.parameter.username : username,
-                      Definition.service.api.parameter.password : password,
-                      Definition.service.api.parameter.id_spot : spotId]
-        }
-        let api = Definition.service.api.remove_f_spot
-        let url = Definition.service.url(api: api, tokens: tokens)
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+                            password: String?) async throws -> WGSuccess? {
+        try await request(tokens: [Definition.service.api.parameter.id_spot: spotId,
+                                   Definition.service.api.parameter.username: username,
+                                   Definition.service.api.parameter.password: password],
+                          api: .remove_f_spot)
     }
     
     // spots = api(query: routine.spots,
@@ -281,105 +159,39 @@ public extension ForecastWindguruService {
     //                     parameter.id_country,       // "id_country" (optional)
     //                     parameter.id_region,        // "id_region" (optional)
     //                     parameter.opt])             // opt=simple (optional)
-    
-    
     func spots(withCountryId countryId: String?,
-               regionId: String?,
-               failure: @escaping FailureType,
-               success: @escaping (_ spotResult: SpotResult?) -> Void) {
-        
-        var tokens = [Definition.service.api.parameter.with_spots : "1"]
-        if let countryId = countryId {
-            tokens[Definition.service.api.parameter.id_country] = countryId
-        }
-        if let regionId = regionId {
-            tokens[Definition.service.api.parameter.id_region] = regionId
-        }
-        let api = Definition.service.api.spots
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+               regionId: String?) async throws -> SpotResult? {
+        try await request(tokens: [Definition.service.api.parameter.with_spots: "1",
+                                   Definition.service.api.parameter.id_country: countryId,
+                                   Definition.service.api.parameter.id_region: regionId],
+                          api: .spots)
     }
-    
     
     // modelInfo = api(query: routine.model_info,
     //             errorCode: err.model_info.rawValue,
     //            parameters: [parameter.id_model])    // id_model (optional)
-    
-    func modelInfo(onlyModelId modelId: String?,
-                   failure: @escaping FailureType,
-                   success: @escaping (_ model: Models?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let modelId = modelId {
-            tokens[Definition.service.api.parameter.id_model] = modelId
-        }
-        let api = Definition.service.api.modelInfo
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+    func modelInfo(onlyModelId modelId: String?) async throws -> Models? {
+        try await request(tokens: [Definition.service.api.parameter.id_model: modelId],
+                          api: .modelInfo)
     }
+    
     
     // modelsLatLon = api(query: routine.models_latlon,
     //                errorCode: err.models_latlon.rawValue,
     //               parameters: [parameter.lat, parameter.lon])
     func models(bylat lat: String?,
-                lon: String?,
-                failure: @escaping FailureType,
-                success: @escaping (_ models: [String]?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let lat = lat, lat.count > 0,
-           let lon = lon, lon.count > 0 {
-            tokens[Definition.service.api.parameter.lat] = lat
-            tokens[Definition.service.api.parameter.lon] = lon
-        }
-        let api = Definition.service.api.modelsLatLon
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponseWithString(response,
-                                           url: url, api: api,
-                                           context: "\(#file):\(#line):\(#column):\(#function)",
-                                           failure: failure, success: success)
-        }
+                lon: String?) async throws -> String? {
+//        [""]
+        try await request(tokens: [Definition.service.api.parameter.lat: lat,
+                                   Definition.service.api.parameter.lon: lon],
+                          api: .modelsLatLon)
     }
-    
     
     // geoRegions = api(query: routine.geo_regions,
     //              errorCode: err.geo_regions.rawValue,
     //             parameters: nil)
-    
-    func geoRegions(withFailure failure: @escaping FailureType,
-                    success: @escaping (_ model: GeoRegions?) -> Void) {
-        
-        let api = Definition.service.api.geoRegions
-        let url = Definition.service.url(api: api, tokens: [:])
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+    func geoRegions() async throws -> GeoRegions? {
+        try await request(tokens: [:], api: .geoRegions)
     }
     
     // countries = api(query: routine.countries,
@@ -387,25 +199,9 @@ public extension ForecastWindguruService {
     //            parameters: [parameter.with_spots,        // with_spots = 1 (optional)
     //                         parameter.id_georegion])     // "id_georegion" (optional)
 
-    func countries(byRegionId regionId: String?,
-                   failure: @escaping FailureType,
-                   success: @escaping (_ countries: Countries?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let regionId = regionId {
-            tokens[Definition.service.api.parameter.id_georegion] = regionId
-        }
-        let api = Definition.service.api.countries
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+    func countries(byRegionId regionId: String?) async throws -> Countries? {
+        try await request(tokens: [Definition.service.api.parameter.id_georegion: regionId],
+                          api: .countries)
     }
     
     // regions = api(query: routine.regions,
@@ -413,121 +209,37 @@ public extension ForecastWindguruService {
     //          parameters: [parameter.with_spots,     // with_spots = 1 (optional)
     //                       parameter.id_country])    // "id_country" (optional)
     
-    func regions(byCountryId countryId: String?,
-                 failure: @escaping FailureType,
-                 success: @escaping (_ regions: Regions?) -> Void) {
-        
-        var tokens : [String: String?] = [:]
-        if let countryId = countryId {
-            tokens[Definition.service.api.parameter.id_country] = countryId
-        }
-        let api = Definition.service.api.regions
-        let url = Definition.service.url(api: api, tokens: tokens)
-        
-        Communication.request(url) {
-            [weak self]
-            (response) in
-            self?.mapAndResponse(response,
-                                 url: url, api: api,
-                                 context: "\(#file):\(#line):\(#column):\(#function)",
-                                 failure: failure, success: success)
-        }
+    func regions(byCountryId countryId: String?) async throws -> Regions? {
+        try await request(tokens: [Definition.service.api.parameter.id_country: countryId],
+                          api: .regions)
     }
 }
 
 private extension ForecastWindguruService {
-    func mapAndResponse<T: BaseMappable>(_ response: Response?,
-                        url: String,
-                        api: ForecastWindguruService.Definition.service.api,
-                        context: String,
-                        failure:@escaping (_ error: WGError?) -> Void,
-                        success:@escaping (_ result: T?) -> Void)
-    {
-        var error : Error? = nil
-        var data : Data? = nil
-        var json : [String:Any]? = nil
-        if let response = response {
-            data = response.data
-            error = response.error
-            do {
-                if let data = data {
-                    json = try JSONSerialization.jsonObject(with: data, options:[]) as? [String : Any]
-                    if let json = json,
-                        WGError.isMappable(map: json) == true
-                    {
-                        DispatchQueue.main.async(execute: {
-                            failure(WGError.init(map: json))
-                        })
-                        return
-                    }
-                }
-            }
-            catch {
-                
-            }
+    
+    func request(tokens: [String: String?],
+                 api: ForecastWindguruService.Definition.service.api) async throws -> String? {
+        let url = Definition.service.url(api: api, tokens: tokens)
+        let (data, _) = try await URLSession.shared.data(from: URL.init(string: url)!)
+        if let string = String.init(data: data, encoding: .utf8) {
+            print("SUCCESS url = \(url) - response.result.value \(string)")
+            return string
+        } else {
+            print("FAILURE url = \(url)")
+            throw CustomError.invalidParsing
         }
-        
-        if  let json = json,
-            error == nil,
-            let klass = T.self as? Mappable.Type, // Check if object is Mappable
-            let object = klass.init(map: json) as? T
-        {
-            print("SUCCESS url = \(url) - response.result.value \(String(describing: json))")
-            DispatchQueue.main.async(execute: {
-                success(object)
-            })
-        }
-        else {
-            print("FAILURE url = \(url) - response.result.error \(String(describing: error))")
-            DispatchQueue.main.async(execute: {
-                failure(WGError.init(code: api.errorCode,
-                                     desc: "failed using=\(url).",
-                    reason: "\(api.query) failed",
-                    suggestion: "\(#file):\(#line):\(#column):\(#function)",
-                    underError: error as NSError?,
-                    wgdata: data))
-            })
-        }
-        
     }
 
-    func mapAndResponseWithString(_ response: Response?,
-                                  url: String,
-                                  api: ForecastWindguruService.Definition.service.api,
-                                  context: String,
-                                  failure:@escaping (_ error: WGError?) -> Void,
-                                  success:@escaping (_ result: [String]) -> Void)
-    {
-        var error : Error? = nil
-        var data : Data? = nil
-        if let response = response {
-            data = response.data
-            error = response.error
-        }
-        
-        if  let response = response,
-            response.error == nil,
-            let responseData = response.data,
-            let jsonString = String(data: responseData, encoding: .utf8)
-        {
-            let array = Array(Set(jsonString.localizedLowercase.replacingOccurrences(of: "]", with: "").replacingOccurrences(of: "[", with: "").replacingOccurrences(of: "\"", with: "").components(separatedBy: ","))).sorted()
-            
-            print("SUCCESS url = \(url) - response.result.value \(String(describing: jsonString))")
-            DispatchQueue.main.async(execute: {
-                success(array)
-            })
-        }
-        else {
-            print("FAILURE url = \(url) - response.result.error \(String(describing: error))")
-            DispatchQueue.main.async(execute: {
-                failure(WGError.init(code: api.errorCode,
-                                 desc: "failed using=\(url).",
-                reason: "\(api.query) failed",
-                suggestion: context,
-                underError: error as NSError?,
-                wgdata: data))
-            })
+    func request<T: Mappable>(tokens: [String: String?],
+                              api: ForecastWindguruService.Definition.service.api) async throws -> T? {
+        let url = Definition.service.url(api: api, tokens: tokens)
+        let (data, _) = try await URLSession.shared.data(from: URL.init(string: url)!)
+        if let json = try JSONSerialization.jsonObject(with: data, options:[]) as? [String : Any] {
+            print("SUCCESS url = \(url) - response.result.value \(String(describing: json))")
+            return T.init(map: json)
+        } else {
+            print("FAILURE url = \(url)")
+            throw CustomError.invalidParsing
         }
     }
 }
-*/
